完成topk的初步修缮，重新设计了前向的复杂度估计方式，rmax仍是两者相等的时候

+++ mylib.h
@@ -651,7 +651,7 @@ static string replace(std::string str, const std::string &from, const std::strin
 }
 
 // unit is kilobyte (KB)
-inline int get_proc_memory(){
+inline long long get_proc_memory(){
     struct rusage r_usage;
     getrusage(RUSAGE_SELF,&r_usage);
     return r_usage.ru_maxrss;
